Did some stuff... Can't remember what.
I know I fixed bug regarding field accesses.
That was fun...
I forgot that a.b should desugar to a["b"], but I incorrectly
desugared it to a[b].
I fully removed the variable AST thing, as I believe it was a terrible solution.
And not even a complete one at that.
I'm still thinking about how I'm going to solve the assignment problem;
I might have to restructure the assignment ASTNode.