Introduce Name node, a semantic analysis phase, and more.
The Name ast node was introduced to handle all variables within the language.
UPVALs still aren't supported, as I need to get a couple of things sorted out first, such as how to store variables on the stack, as I don't think how I do it currently will work well with that.
Also, UPVALs within the old implementation work a bit weirdly, so I want to avoid that problem.
With the Name node comes some cleanup of the structure itself, along with a more defined border between node uses. It's a lot easier to spot known local variables.
(eg, parameters are marked as locals, because they're not going to bind another value)

Assignment usage was also cleaned up. It's easier to see what is actually a language level assignment, rather than just a desugared statement.