Goto, break, and labels now work. (As a result, for loops now work.)
After implementing the control flow stuff, the for loops should have worked correctly right out of the gate, but didn't...
There were some lovely problems.
First, during the construction of the false jump list, I forgot to concat the relevant lists when constructing a sub-cond with the AND and OR emitter code.
Second, the get code from the ByteCodeWriter was terrible, and it was coearcing the byte to a full int value. Bitmask solved it, but I'm not happy with the result.
Third, while trying to nest them, I realised I was generating the loops with the same variable names, so it was just ruining everything.
The name that the loops use is now dependent on the block's system hashcode.